+++ harjus-port/engine.h
@@ -19,4 +19,4 @@ private:
   QMultiHash<QString, PlannedExecution> mSymbolToPlannedExecutions;
   QHash<QString, Offer> mSymbolToOffer;
 };
-double get_double_from_json(const QJsonObject &json, const QString &key);
+long double get_double_from_json(const QJsonObject &json, const QString &key);
